+++ Server/src/API/setupAPI.ts
@@ -1,11 +1,11 @@
 import { Express } from 'express-serve-static-core';
 import { setupUpload } from './questions.js';
 import { setupServerStatus } from './serverStatus.js';
-import { setupGame } from './game.js';
+import { setupTest } from './setupTest.js';
 
 // IMPORTA TUTTE LE API DEI VARI FILE
 export const setupAPI = (app: Express) => {
   setupUpload(app);
   setupServerStatus(app);
-  setupGame(app);
+  setupTest(app);
 }
